requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=9.0.0
imagehash>=4.3.0
boto3
minio
matplotlib
glob2
tqdm
imagededup
pytest
icecream